job:#4896 I added an attribute, distance, to class Element Visibility (PE_VIS) that will hold the "relative distance" from the starting point of the search to the specific result in the search result.
I added a parameter, integer distance, to all the places that have an
operation, collectVisibleElementsForName. This was: Component, System
Model, Package, Packageable Element. 

In Package::collectVisibleElementsForName I look to see if we are
"ascending" in the model of "descending" and I increment/decrement the
new distance attribute accordingly before passing it along to recursive
calls to collectVisibleElementsForName.

There were a few places, where a search is initiated. In those places,
the value of distance starts at 0.  This means as we move "up and down"
the model hierarchy the value of "distance" may go from negative to
positive while the search finds all the instances..