job:#10004 Do not allow OAL parsing unless the OAL dialect is selected
Introduced class "BridgePointDialectException extends Exception" and
modified the bp.als.oal/TextParser.java constructors to throw this
exception if they are invoked when the OAL dialect is not selected. The
message for this exception in TextParser.java is: "The OAL dialect is
not selected, OAL parsing is not allowed."

With the above change in place I then built the code to find the places
where this new exception needed to be handled and I modified these
places to catch the exception, report the error, and return gracefully.
To assure I did not miss any usage locations I temporarily commented-out
the TextParser.java constructors and built again and checked to assure
all usage places had indeed been handled. I also introduced a private
TextParser.java no-arg constructor to assure that was no a possibility
for usage.