job:#10005 Use "_" instead of "-" when finding a unique name
During copy/paste (and other places) the tool sometime has a need to
assure a unique name is present in the file system. In doing this it
adds a incrementing number to a name candidate. I changed the separator
used between the base name and the prefix from "-" to "_" to prevent
potential problems with dialects that don't allow "-" in names.